Update metrics tests
